Compile without warnings on VS2022
Removed hundreds of warnings. Mostly casting CString to LPCTSTR
and casting LONGs to LONG_PTR, DWORDs to DWORD_PTR, and so on.
And tr1 is finally proper std in Visual Studio.

Some unused code has been removed:
* Loading res-dll in InitInstance() had a lot of broken code. Works well now.
* MFCExt had some unused classes.
The code has been removed since it caused crashes or warnings.
There may still be more code that is unused.

+++ MFCExt/Source/IniFile.cpp
@@ -87,12 +87,12 @@ bool CIniFile::ReadFile()
 		return true;
 
 	std::vector<char> data(s);
-	file.Read(&data[0],s);
+	file.Read(&data[0], (DWORD)s);
 
-	int cch = ::MultiByteToWideChar(CP_UTF8,0,&data[0],s,0,0);
+	int cch = ::MultiByteToWideChar(CP_UTF8, 0, &data[0], (DWORD)s, 0, 0);
 
 	std::vector<wchar_t> buffer(cch + 1);
-	::MultiByteToWideChar(CP_UTF8,0,&data[0],s,&buffer[0],cch + 1);
+	::MultiByteToWideChar(CP_UTF8, 0, &data[0], (DWORD)s, &buffer[0], cch + 1);
 
 	std::basic_istringstream<TCHAR> inifile(&buffer[0]);
